d_a_obj_try.h: REVERT m64C/m64D u8->s8 again (D7's b66426602 re-applied u8, breaking d_a_obj_correct Matching rel AGAIN via correct.h->try.h include; retail correct pin bee1e391 requires s8 header) + try.cpp keeps (u8) casts at all 3 compare sites for lbz+cmplwi semantics; mode_wait/sink_init unaffected

+++ src/d/actor/d_a_obj_try.cpp
@@ -593,7 +593,7 @@ void daObjTry::Act_c::mode_wait() {
     }
 
     int ret = 0;
-    if (groundHit && (bingo != 0 || m64D != 0)) {
+    if (groundHit && (bingo != 0 || (u8)m64D != 0)) {
         ret = 1;
     }
